% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compression_methods}
\alias{compression_methods}
\title{List of all supported compression methods}
\usage{
compression_methods()
}
\value{
A character vector of method identifiers.
}
\description{
Returns the methods accepted by the \code{method} argument of
\code{\link[=compress_posterior]{compress_posterior()}} and friends.
}
\examples{
compression_methods()
}
