% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan-utils.R
\docType{data}
\name{poco_stan_models_cache_dir}
\alias{poco_stan_models_cache_dir}
\title{Default cache directory for compiled Stan models}
\format{
An object of class \code{character} of length 1.
}
\usage{
poco_stan_models_cache_dir
}
\value{
A character string with the path to the cache directory.
}
\description{
Mirrors \code{sccomp::sccomp_stan_models_cache_dir}. Used as the default cache
location when \code{poco} (or downstream packages built on top of it) compile or
load Stan models via \code{cmdstanr} / \code{instantiate}.
}
\examples{
poco_stan_models_cache_dir
}
\keyword{datasets}
