% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_density.R
\name{print.posterior_compressed}
\alias{print.posterior_compressed}
\title{Print method for compressed posteriors}
\usage{
\method{print}{posterior_compressed}(x, ...)
}
\arguments{
\item{x}{A \code{posterior_compressed} object.}

\item{...}{Currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for compressed posteriors
}
