% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct_sccomp}
\alias{reconstruct_sccomp}
\title{Reconstruct an sccomp object from a compressed posterior}
\usage{
reconstruct_sccomp(x, n_draws = NULL)
}
\arguments{
\item{x}{The output of \code{\link[=compress_sccomp]{compress_sccomp()}}.}

\item{n_draws}{Integer number of samples to regenerate. If \code{NULL},
uses the original number of draws.}
}
\value{
The sccomp object with \code{attr(x, "fit_compressed")} populated.
}
\description{
The compressed posterior and regenerated samples are stored in
\code{attr(x, "fit_compressed")}. The original \code{attr(x, "fit")} is removed.
}
