% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_density.R
\name{summary.posterior_compressed}
\alias{summary.posterior_compressed}
\title{Summary method for compressed posteriors}
\usage{
\method{summary}{posterior_compressed}(object, ...)
}
\arguments{
\item{object}{A \code{posterior_compressed} object.}

\item{...}{Currently unused.}
}
\value{
A list with class \code{summary.posterior_compressed}.
}
\description{
Summary method for compressed posteriors
}
